compress: Get LZ4 working on Windows
For LZ4 to work on Windows, you have to define an environment variable
called LZ4_DIR pointing to your LZ4 install.  Once done, these
modifications make it possible to find the includes and libraries
appropriately on WIndows.

+++ util/lz4_compressor.cpp
@@ -46,8 +46,8 @@ size_t Lz4Compressor::Compress(const size_t          uncompressed_size,
                 LZ4_compress_fast_continue(&lz4_stream,
                                            reinterpret_cast<const char*>(uncompressed_data),
                                            reinterpret_cast<char*>(compressed_data->data()),
-                                           uncompressed_size,
-                                           lz4_compressed_size,
+                                           static_cast<const int32_t>(uncompressed_size),
+                                           static_cast<int32_t>(lz4_compressed_size),
                                            1);
             if (compressed_size_generated > 0)
             {
@@ -80,8 +80,8 @@ size_t Lz4Compressor::Decompress(const size_t                compressed_size,
                 LZ4_decompress_safe_continue(&lz4_stream_decode,
                                              reinterpret_cast<const char*>(compressed_data.data()),
                                              reinterpret_cast<char*>(uncompressed_data->data()),
-                                             compressed_size,
-                                             expected_uncompressed_size);
+                                             static_cast<int32_t>(compressed_size),
+                                             static_cast<int32_t>(expected_uncompressed_size));
             if (uncompressed_size_generated > 0)
             {
                 copy_size = uncompressed_size_generated;
